chore: refactored package name to domain
