+++ src/domain/Axis.java
@@ -1,4 +1,4 @@
-package core;
+package domain;
 
 public class Axis {
