+++ src/domain/Coordinate.java
@@ -1,3 +1,3 @@
-package core;
+package domain;
 
 public record Coordinate (int x, int y) {}
