+++ src/domain/Helicopter.java
@@ -1,4 +1,4 @@
-package core;
+package domain;
 
 import java.util.HashMap;
 import java.util.Map;
