+++ src/domain/Item.java
@@ -1,4 +1,4 @@
-package core;
+package domain;
 
 public interface Item {
