+++ test/test/LocationMapperTest.java
@@ -2,7 +2,7 @@ package test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import core.Location;
+import domain.Location;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
